+++ community/cypher/cypher/src/main/scala/org/neo4j/cypher/internal/compatibility/OnlyOnceQueryExecutionMonitor.scala
@@ -19,7 +19,7 @@
  */
 package org.neo4j.cypher.internal.compatibility
 
-import org.neo4j.kernel.api.ExecutingQuery
+import org.neo4j.kernel.api.query.ExecutingQuery
 import org.neo4j.kernel.impl.query.QueryExecutionMonitor
 
 case class OnlyOnceQueryExecutionMonitor(monitor: QueryExecutionMonitor) extends QueryExecutionMonitor {
