+++ community/kernel/src/main/java/org/neo4j/kernel/api/query/PlannerInfo.java
@@ -23,8 +23,6 @@ import java.util.List;
 
 import javax.annotation.Nonnull;
 
-import org.neo4j.kernel.api.index.IndexUsage;
-
 public class PlannerInfo
 {
     private final String planner;
