+++ community/kernel/src/main/java/org/neo4j/kernel/impl/api/operations/QueryRegistrationOperations.java
@@ -22,7 +22,7 @@ package org.neo4j.kernel.impl.api.operations;
 import java.util.Map;
 import java.util.stream.Stream;
 
-import org.neo4j.kernel.api.ExecutingQuery;
+import org.neo4j.kernel.api.query.ExecutingQuery;
 import org.neo4j.kernel.impl.api.KernelStatement;
 import org.neo4j.kernel.impl.query.clientconnection.ClientConnectionInfo;
