+++ community/kernel/src/test/java/org/neo4j/kernel/api/query/ExecutingQueryTest.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.kernel.api;
+package org.neo4j.kernel.api.query;
 
 import java.time.ZonedDateTime;
 import java.util.Arrays;
@@ -31,6 +31,7 @@ import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
 import org.junit.Test;
 
+import org.neo4j.kernel.api.query.ExecutingQuery;
 import org.neo4j.kernel.api.query.PlannerInfo;
 import org.neo4j.kernel.impl.locking.LockWaitEvent;
 import org.neo4j.kernel.impl.query.clientconnection.ClientConnectionInfo;
